debian: Detect if we are using RedHat or Debian
Debian uses dpkg to manage packages so using RPM
on a Debian based system will not detect the correct
packages and cause the playbook to fail.

To get around this, before the playbook is run we
turn on fact gathering and determine if we are running
either RedHat (Centos 7) or Debian (Debian bullseye).

Test:
- Applied patch
- Ran playbook with applied patch to see if the detection
  works properly.
- Ran playbook sucessfully on both Debian and Centos.

Tested successfully on both Debian and Centos.

Story: 2009101
Task: 44158

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: Ia3a42c305e56c2f7e6537b7f44a6ed41869b4417