Playbook fails at alarm check task
This commit adds retry for the alarm check in order to make it more
resilient to temporary alarms.

It also includes some minor improvements regarding certificate backups.

Test Plan:
PASS: Trigger temporary alarm and run the playbook. It should be
      resilient enough to wait for about 2 minutes for alarm to clear.
PASS: Trigger a persistent alarm and check that playbook waits for
      about 2 minutes for alarm to clear and fails when it does not.

Closes-Bug: 1958885
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I44d3a885d0f470263fd9feb8947755b4cb67326b