Install Root CA when migrating certificates
This is a refactor of the ssl_ca certificate install in order to
support multiple certificates by specifying certificates as parameters.

Now, the playbook also installs the Root CA certificate alongside the
ICA certificate. That fixes the bug with system registry-image-list.

Also, updated naming for some parameters and documentation in
migrate-subcloud1-overrides-EXAMPLE.yml to better explain parameters
now that Root CA is introduced.

Test Plan:
PASS: Run the playbook and system certificate-list. Check that both
      system_local_ca_cert and root_ca_cert are installed.
PASS: Run the playbook and check that system registry-list returns with
      no errors
PASS: Run the playbook with unrelated system_local_ca_cert and
      root_ca_cert and verify the playbook shows an error message

Closes-Bug: 1958932

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: Iff0dfd51fcc8e4bdb06c13789e079423a27a8cb8