Add a new playbook to prestage software packages
This commit adds a new playbook to prestage software packages
on the target in preparation for upgrade or reinstall. The
playbook is meant to be integrated with dcmanager.

Fixes for images prestaging where the target does not have docker proxy
configured and for some lint warnings are also included in the commit.

Usage:
  ansible-playbook prestage_sw_packages.yml \
      -i <inventory-file> \
      -e "software_version=<release-number> \
          shared_download_source=<download-source-dir>"

For testability, an optional parameter shared_download_source is
added to the playbook. If not specified in the extra-vars list, the
playbook will look for the required content at
/opt/platform/deploy/<software_version>/prestage/shared directory on
the system controller.

Password used for privilege escalation (ansible_become_pass) must be
provided either as part of the inventory file or command line.

Test plan:
  Success cases:
    - Prestage software packages for reinstall (i.e. specified
      software version is the same as on the target)
    - Prestage software packages for upgrade (i.e. specified
      software version is higher than on the target)
    - Playbook replay, completed

  Failure cases:
    - Prestage packages with invalid software version
    - Packages comps.xml file is missing
    - Any of other Packages download sources is missing

Story: 2009799
Task: 44464
Change-Id: I3ac70a4f2fb2417118aeb3844ca3759470ca84c3
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>