Using general openstack app name when restoring it
As part of the work being done to reduce the friction of making a
downstream release of stx-openstack, the Openstack application is being
handled the most "app name agnostic" as possible.

This change aims to add a new variable 'openstack_app_name' to store the
Openstack application name and to be referred every time the
backup/restore playbooks need to operate over that app.
Its default value will still be 'stx-openstack' to keep the default
behaviour as it is now, but it will be possible to pass this value as an
argument when calling playbooks/backup.yml (i.e -e
"openstack_app_name=downstream-name")

TEST PLAN:
PASS: Build & Install on an AIO-SX system
PASS: Execute stx-openstack application backup procedure [1]
successfully
PASS: Execute stx-openstack application restore procedure [1]
successfully

[1] https://docs.starlingx.io/developer_resources/backup_restore.html

Story: 2009669
Task: 44540

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I95a45a5c2d3abef59f3a695fc979108d73dc8955