Add mode to delete LDAP users
Create a new parameter mode to remove LDAP user with keystone account
and sudo access if exists. The playbook run a set of tasks to remove
the user in all subclouds.

The objective of the playbook is to remove the specified ldap account
to disallow future ssh connects to the system using the account.
The playbook works for stand-alone as well as Distributed Cloud
systems.

Test Plan:
  PASS: Create user with mode 'create' and then delete the
        same user using mode 'delete'.
  PASS: Verify the user deleted can not connect using SSH to all
        subclouds
  PASS: Verify the script can run in stand-alone system
  PASS: Verify the script can run in distributed cloud

Story: 2009759
Task: 44537

Signed-off-by: Alexandre Horst <alexandre.horst@windriver.com>
Change-Id: I7271d77b2daa5beb5d55052ad7c9c2f0c4f36719