Increase filesystems size for systemcontroller
On a DC environment, it is needed to increase platform and
backup filesystems size to improve scalability on parallel
subclouds deployment (100+ deployments in parallel).

This commit increases both filesystems for setups with
a systemcontroller role and disk sizes greater than the
small disk size (ONLY for standard deployments, that allocate
all the remaining disk space to cgts-vg [1], as opposed to AIO
that allocates a fixed size [2]).

Limitation: this commit changes the apply_bootstrap_manifest
tasks file, which is not executed on replay, so, the changes
present on this commit will only apply during the first bootstrap
on the host. In a scenario where replay bootstrap is executed to
change distributed_cloud_role the user will have to resize the
backup and platform filesystems via system host-fs-modify and
controllerfs-modify commands manually.

[1] https://opendev.org/starlingx/metal/src/branch/master/bsp-files/kickstarts/pre_disk_controller.cfg#L75
[2] https://opendev.org/starlingx/metal/src/branch/master/bsp-files/kickstarts/pre_disk_aio.cfg#L177

Test Plan:
PASS: install and bootstrap AIO-SX, AIO-DX and Standard
      non-System Controller and verify that following filesystems
      are created with sizes:
      backup=25G, platform=10G
PASS: install and bootstrap AIO-DX System Controller and verify
      that following filesystems are created with sizes:
      backup=25G, platform=10G
PASS: install and bootstrap Standard System Controller and verify
      that following filesystems are created with sizes:
      backup=35G, platform=20G
PASS: bootstrap a system with distributed_cloud_role=none, change
      the localhost.yml and set distributed_cloud_role to
      systemcontroller, run bootstrap again and verify that the
      distributed_cloud_role changes, but filesystems size remain
      unchanged on the host filesystem and in the database

Story: 2009725
Task: 45208
Change-Id: I0a14ee25aa6d2d9c96f172a45ac81cfa25aa6eb3
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>