Improve known_hosts and password update
This story/task is for resolving the random failures
when executing"Add remote SSH public keys into the known_hosts"
and "Change initial password" tasks.

There are two potential issues in adding SSH public keys into
the known_hosts. The first is the "lineinfile" module does not
support concurrency properly. Sometimes the ssh key was not
really added to known_hosts although the task succeeded, so we
added check and retries for it. The second issue derived from
its previous task "Gather remote SSH publc key". The ssh-keyscan
might hang there for a long time (more than 5s which is the
default timeout value for this command) without any result.
Then the adding ssh key task would be silently ignored since
the sshkey scan result is empty. For this case, we increased
the timeout value and added retries to the ssh-keyscan task.

In terms of the failure in "Change initial password", the default
60s timeout is not enough, so we increased the timeout value.
Besides, we fixed the rescue. The previous rescue didn't match
the failure scenario.

Test Plan:
  - PASS: Install/reinstall a large number of subclouds in a batch
          and verify there are no Authentication failures.

Story: 2009725
Task: 44839

Change-Id: I1429e8b6cf8f27b3949b6eabc78843c7aed20cf3
Signed-off-by: Li Zhu <li.zhu@windriver.com>