Fetch docker lv_size from backup hieradata
Currently the bootstrap recovers only a subset of hieradata
from the backup file, namely the static and system yaml files.
This, in turn, allows the bootstrap to configure the drbd
filesystem sizes accordingly to the values configured by the
user before the upgrade.

This commit adds additional steps on bootstrap playbook to
get docker lv_size from the backup tarball and include it
on the runtime.yaml so that docker lv_size is configured
during the bootstrap with the custom value provided by the
user.

Test Plan:
PASS: 1. Increase docker filesystem prior to upgrade
      2. Upgrade AIO SX and verify that during bootstrap the
         docker filesystem is configured with custom value
         configured prior to the upgrade
PASS: 1. Increase subcloud docker filesystem prior to upgrade
      2. Upgrade subcloud and verify that during bootstrap the
         docker filesystem is configured with custom value
         configured prior to the upgrade

Closes-bug: 1968779
Change-Id: Icbf5cf8773ce09b8ec830571167efe26e64f51fe
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>