debian: Update NO_PROXY for docker
This work enables bootstrap to pass for IPv6 with docker proxies and
docker no_proxy deployments.

On Debian we use a newer version 20.10 of docker vs 18.09 on CentOS.
We require to use no square brackets IPv6 notation for NO_PROXY
variable.
Same behavior was observed for containerd in the past[1].
[1]: https://review.opendev.org/c/starlingx/ansible-playbooks/+/704677

Tests AIO-SX IPv6 Debian:
PASS: build-pkgs && build-image
PASS: bootstrap on IPv6 with proxies and no_proxy configured

Skip CentOS tests because logic on CentOS was not changed, overall logic
was split based on os_release variable which is validated in other
places.

Story: 2009964
Task: 45112
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Ic683a5e889874c778dfd5103e488eaf58c74c654