Added default audit log parameters and volumes
Default values for parameters "audit-log-path", "audit-log-maxage",
"audit-log-maxbackup" and "audit-log-maxsize" were added to
kube-apiserver startup. Also, the volumes for log folder and default
audit policy file were added in kube-apiserver pods.

The kubernetes audit is not enabled by default, as the parameter
"audit-policy-file" is not set at kube-apiserver startup. However,
it is ready to be enabled with the parameter "audit-policy-file"
pointing to the default audit policy file path.

Test Plan:

PASS: Successfully deploy an AIO-DX using an image with this commit
present.
PASS: In the deployed AIO-DX, verify with command "kubectl describe pod"
that the pods "kube-apiserver-controller-0" and
"kube-apiserver-controller-1" are running with default values for
parameters "audit-log-path", "audit-log-maxage", "audit-log-maxbackup"
and "audit-log-maxsize" and that there are volumes for log folder and
default audit policy file.
PASS: In the deployed AIO-DX, in the active controller, enable
kubernetes audit by adding the parameter
"--audit-policy-file=/etc/kubernetes/default-audit-policy.yaml" in the
manifest "/etc/kubernetes/manifests/kube-apiserver.yaml" and check that
it is working correctly.
PASS: In the deployed AIO-DX, swact to the other controller and repeat
the test above (enabling and checking the feature).

Story: 2009835
Task: 45202

Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I2e1dd8be6f783194fe2f7b660845ad3cd21d3b70