Fix ldap dir on ansible playbook for backup roles
when running backup on Debian, ansible throws an error
on ldap db backup task.
error: "slapcat: 'bad configuration directory!'"

that's because ldap configuration directory changed on
debian from openldap to ldap

the fix this is to indentify the os distribution that
is currently using and do a conditional on ansible, to
differentiate the task on debian from centos, to be able
to fetch information from the correct directory depending
on the os.

Test Plan:
PASS: Build an image on Debian
PASS: Build an image on CentOS
PASS: Create backup on Debian
PASS: Create backup on CentOS

Story: 2009303
Task: 45477

Signed-off-by: Ramon Gazoni Lacerda <Ramon.GazoniLacerda@windriver.com>
Co-Authored-By: Daniel Stevens <DanielStevens.TorresCardenas@windriver.com>
Co-Authored-By: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I84ea3ebb3c0f31c03d678c02fee79fc298492313