Revert "Use script to update kernel parameters for Debian"
This reverts commit b28436a3eef0c8272dc22fb6c092188aadbd6204.

Reason for revert: 
Since commit 0f6eb74ed7415defe94dc8b92c5cb9c16aee8b26 was 
reverted, this commit also needs to be reverted.
On a hardware lab there is a puppet failure
due to incorrectly handing the hugepage parameter on the command
line.  This causes the controller to fail to go enabled and a
reboot loop.
The original commit can be re-posted for review once this issue
is fixed.

Change-Id: I59ea385153ae08d5d5795e9b4976ce0b86e158ef