Add CNI dir config to multus plugin
This commit adds the configuration of the cniDir, confDir and binDir
to the configmap of multus-cni. This is required because application
can have interfaces added by Multus using istio-cni when istio is
running. istio-cni is installed under /opt/cni/bin by IstioOperator.

Test Plan:
Pass: Ensure that the multus-cni-config configmap is updated
Pass: Regression of Istio application

Story: 2009912
Task: 45515

Change-Id: I902d0138ee9ee7563cd2df3d4b9de535933a1847
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>