Stop platform pods from requesting CPU resources
While running tests in the lab, calico-node and
kube-sriov-device-plugin pods are requesting CPU resources.
All platform pods are supposed to have no CPU resource requests
as it messes up the reported CPU allocations and reduces the
amount of CPU available to the customer applications

The CPU request allocation for calico-node and
kube-sriov-device-plugin containers have been changed
to "0". This will implement no CPU requests for pods
in the platform namespaces

The kube-sriov-device-plugin pod CPU requests was modified in
a SR-IOV enabled lab, to help verify that the code
change was successful

Test Plan:

PASS: set calico-node requests to 0 local build VM
      the following command was executed:
      "kubectl describe node controller-0"
      --> command response:
          Non-terminated Pods:
          .......
          calico-node-kxbpl             CPU requests: 0 (0%)

PASS: set sriovdp-node requests to 0 on SR-IOV enabled lab
      the following command was executed:
      "kubectl describe node controller-0"
      --> command response:
          Non-terminated Pods:
          .......
          kube-sriov-device-plugin-amd64-p5s4g   CPU requests: 0 (0%)

Closes-Bug: 1976300

Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>
Change-Id: I591886e3dd14cbbef0e0743707eb4cb44dff70a3