Fix active controller check on backup
The current backup ansible code is using the raw "system host-show"
output to grep by "Controller-Active", however, this output is
formatted in a given way that it is prone to have line breaks.
As the field containing this info is open to have other infos, and,
recently, the max_cpu_config was added, this caused a line break in
the output (the field is the "capabilities").
Therefore, as this info is not in the database, since it is added
"on-the-fly" by sysinv, and the system command has the option
to format the output differently, using: "--column capabilities
--format value", the current logic can be kept since the output
comes as a whole string with the field value.

Test Plan:

PASS: Perform backup & restore on AIO-SX successfully

Closes-bug:  1977547

Signed-off-by: Adriano Oliveira <adriano.oliveira@windriver.com>
Change-Id: Iee5bd579558872f8e81624e208af37cd92037485