Move update of resolv.conf from persist-config to validate-config
On python3, when importing eventlet.green library a
verification is done on /etc/resolv.conf to have at least one valid
nameserver. Since we are on a fresh install and the resolv.conf
file is clean on the iso as part of
https://review.opendev.org/c/starlingx/tools/+/844460
It fails the execution

It was attempted to solve the first error in
bootstrap/validate-config/tasks/validate_address.yml with the task
"Check if the supplied address is a valid domain name or ip address"
to isolate the code inside sysinv as can be seen on the abandoned
changes
https://review.opendev.org/c/starlingx/config/+/845003
https://review.opendev.org/c/starlingx/ansible-playbooks/+/845015

But the problem was just postponed to
bootstrap/store-passwd/tasks/main.yml on the task
"Store admin password", and this one uses just python libraries.

So we are just moving the update of resolv.conf from persist-config
to validate-config as this will not change execution times for
bootstrap execution and it works for both CentOS and Debian

With the delivery of
https://review.opendev.org/c/starlingx/config/+/843012 the task
"Remove localhost address from resolv.conf" is no longer needed

Test Plan:
PASS execution of AIO-SX bootstrap for Debian
PASS execution of AIO-SX bootstrap on CentOS

Closes-Bug: 1977887

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I63e9dbaa82eb502d52b6f972c7944857e2547494