Adding SRIOV config.json file restore from backup
During the upgrade process, the kube-sriov-device-plugin-amd64 pod
needs the /etc/pcidp/config.json file which is not restored from backup
during the upgrade process and is only created on the next unlock,
causing the pod stays in 'ContainerCreating' state until unlock is
executed.
This change checks if the config.json file is present in the backup
tarball and restores it if the bootstrap process is in restore mode.

Test Plan:

PASS: AIO-SX Upgrade in a SRIOV enabled hardware
PASS: Check /etc/pcidp/config.json file is restored from backup
PASS: Check pod kube-sriov-device-plugin-amd64 in Running state
before the unlock

Closes-Bug: 1978344

Signed-off-by: Luis Eduardo Angelini Marquitti <luis.eduardoangelinimarquitti@windriver.com>
Change-Id: Icd696d958836c0d837409481abdb531969cd366e