Wait for nginx pod and service on restore
While trying to install armada on a restore scenario, helm
may return a connection refused error because the ingress
resource creation triggers a webhook that may not be ready at
that moment. The webhook is related to nginx pod and services.

This commit adds a step to check if nginx pod and service are
ready and, if not, wait up to 60s for them to be up and ready.

Test Plan
PASS: run AIO-SX backup and restore successfully
PASS: run AIO-SX fresh install, bootstrap and unlock successfully
PASS: run AIO-SX upgrade successfully

Closes-bug: 1978899
Change-Id: Iffbfd1b06b61b057fd5b6b9cfcebc1f4fab6cb36
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>