Wait for services to come up
In a restore scenario, during the restore-more-data stage, the
restarted services may come up but still unable to serve requests,
and this, specifically keystone, will cause mtcAgent to fail when
he is executed on the next step, with a Token Request Failed error.

This commit adds a wait time right after services are restarted,
so that services are given enough time to be up and ready to serve
requests.

Test Plan:
PASS: run AIO-SX backup and restore successfully (while monitoring
      keystone ability to serve requests after the service restart
      and observing that it took some time after the process came
      up for it to be able to reply token requests)

Closes-bug: 1979851
Change-Id: I81f199f8e2f9ba2116350908ac54b1070df8027c
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>