Restore backup fixes
The os_release fact has been moved to 'common' since it needs to
be used not just by bootstrap but by backup and restore as well
(and maybe others).

ldap needs to refer to openldap on CentOS and slapd on Debian.

TEST PLAN:
PASS: Restore backups on CentOS
PASS: Restore backups on Debian

Story: 2009303
Task: 45727
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Co-authored-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I6ce3e2886b509b1f70e63953139f26ecd58fce17