Restore /etc/openldap in CentOS
Directory /etc/openldap was not included in the restore process.
All the files that appear inside this directory after
restore come from the iso installation or are generated
based on default configuration.
So, the fix was to include the entire directory in the restore
playbook.
This fix covers only ldap service on CentOS.

Test Plan:

PASSED: test the fix for CentOS
        Backup with openldap configured
        --> /etc/openldap restored according to backup

PASSED: negative test
        Backup without openldap configured
        --> /etc/openldap restored according to backup

PASSED: negative test on Debian
        --> openldap service skipped

Closes-Bug: 1981082
Depends-on: https://review.opendev.org/c/starlingx/ansible-playbooks/+/843430
Signed-off-by: Virginia Martins Perozim <vmartins@windriver.com>
Change-Id: I1d7441f333fd0aa8ae570551ea261e06cb664540