Fixing unquoted backslashes for grep -E
After the merge of [1], the restore playbook fails with a message
saying "We could be wrong, but this one looks like it might be an issue
with missing quotes" at either L243 or L265. This is due to backslash
not properly scaped for yaml syntax.

[1]https://review.opendev.org/c/starlingx/ansible-playbooks/+/848770

TEST PLAN
PASS run restore playbook

Closes-Bug: #1983475
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: Idcb5f5939a2a706afe1da1075397652e55dd37d0