Deleting ic-nginx-ingress-controller at restore
Once k8s comes up after the etcd restore, there is a span of time
(around 20s) that the pod states have not been updated and are reported
as they were at the point in time where the backup was taken. This
returns that the ic-nginx-ingress-ingress-nginx-controller-XXX pod is
"Ready", but it is not... in several instances during my tests, the pod
was restarted 3-10 seconds after the task "Launch Armada with Helm v3"
failed due to not being able to call the webhook. The proposed solution
is to delete the pod preemptively and wait for it to be recreated and
"Ready".

TEST PLAN
PASS restore on virtual AIO-SX (CentOS)

Closes-Bug: #1978899
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I20bec1fbbf809bfcf5d515ef55c6d47ab968dbf3