Restore with ssl_ca certificate as command option
ssl_ca certificate was included as an optional parameter of restore
playbook.
If it is defined in the command line, it will replace the certificate
defined in backup file.
If it is not defined in the command line, the certificate defined in
the backup file will be used.

Test Plan:

PASSED: expired certificate in backup file during restore
        ssl_ca certificate not defined
        bug reproduced

PASSED: expired certificate in backup file during restore
        valid ssl_ca certificate defined
        restore pass with no errors

PASSED: valid certificate in backup file during restore
        valid ssl_ca certificate not defined
        restore pass with no errors

PASSED: test on remote play

Closes-Bug: 1984116
Signed-off-by: Virginia Martins Perozim <vmartins@windriver.com>
Change-Id: I6c1ddca1097fbbdb0acde7df3097287ab59b4c1c