Fixing replacement of /etc/hosts timeouts
During the restore operation, there is a span of time where we end up
without the /etc/hosts file and, when executing the playbooks remotely,
we reach the 12s timeout for command. This commit fixes it by ensuring
that /etc/hosts is available and with the bare minimum entries at all
times and also add a rescue operation to recover the previous /etc/hosts
file in case any task fails.

TEST PLAN
PASS remote play of the restore playbook
PASS local play of the restore playbook
PASS bootstrap newly installed system

Closes-Bug: #1986693
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: Iff8e56478339f660ec66e2d4f7cd8ad000b4d306