Fix format error during Backup and Restore
When last_kube_extra_config_bootstrap.yaml is created, the yaml
is not properly formatted. It was combining the first
entry on the same line as the definition.

"Bad format"
apiserver_extra_args:   audit-log-maxage: '3'
  audit-log-maxbackup: '10'

"Good Format"
apiserver_extra_args:
  audit-log-maxage: '3'
  audit-log-maxbackup: '10'

This was breaking the backup step of backup and restore.

Test Plan:
 Build/Boostrap/Unlock Debian AIO-SX
 Invoke backup operation on Debian AIO-SX (twice)
 Invoke remote backup operation for Debian AIO-SX

Closes-Bug: 1987307
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I6728ea8fd97ef3abbdeaf0eafeda58a2ec6481b6