Prevent backup archive base64 to be shown on log
This change aims to prevent the base64 representation of the backup
archives to be shown on log when running ansible remotely with debug
output. It's not useful, generates a huge log and the log ends up taking
a long time to be opened for debug on any tool.

TEST PLAN
PASS remote backup

Closes-Bug: #1987306
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I960668c911b36ac66a321b0235d1d7cc69dcfb04