Fix AIO-SX upgrade error on nginx wait for pod
This issue seems to be happening because the execution is too fast
and the pod does not exist when the wait command is executed, resulting
in a 'resource not found' error.

The 'kubectl wait' is an experimental feature and there is a very
similar bug reported by the community in [1], where 'wait' fails to
wait for non-existent resources.

This commit addresses the issue with ansible 'retries and delay'.

[1] https://github.com/kubernetes/kubernetes/issues/83242

Test Plan:

PASS: Run aio-sx upgrade and execute upgrade_platform.yml playbook
      with success.
PASS: Fresh install and successful run of the bootstrap playbook.

Closes-Bug: 1989018
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I1b077ac85fa912356ef4e4c9f05b417469296ade