Create openldap certificate on upgrade
This commit adds a playbook to invoke role install-platform-certificates
and create the openldap certificate during upgrade. It has no effect to
any other existing playbook.

Test Plan:

PASS: Upgrade from stx 22.06 to 22.12 and verify that the openldap
      certificate is successfully created in kubernetes
      in upgrade activate stage
PASS: After certificate is created in kubernetes, verify that the
      certificate and key is read by puppet and saved to files
      in /etc/ldap/certs
PASS: After certificate is created in kubernetes, verify that the
      certificate is picked up by cert-mon and saved to database
      and show in 'system certificate-list'

Story: 2009834
Task: 45933

Change-Id: I47d44a1567b952b72002185094e79683ed8a101d
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>