Add user role in LDAP playbook
This commit adds the parameter "user_role" to the playbook
"manage_local_ldap_account.yml", used to create LDAP accounts. This is
an optional parameter that can be set to the values "admin", "member" or
"reader". When not informed in playbook command, it defaults to the
value "admin".

The reference doc for adding LDAP users through playbook is
https://docs.starlingx.io/security/kubernetes/
manage-local-ldap-39fe3a85a528.html .

Test Plan:

PASS: Successfully deploy an AIO-DX using a CentOS image with this
commit present. After, execute the LDAP playbook with the following
variations:
1) Without extra-var "user_role";
2) With extra-var "user_role" using "aaaaaa" as its value;
3) With extra-var "user_role" using "reader" as its value;
4) With extra-var "user_role" using "member" as its value;
5) With extra-var "user_role" using "admin" as its value.
In case 1, the user is created with role "admin"; in case 2, the
playbook execution fails informing that "user_role" value is different
from "reader", "member" or "admin"; in cases 3, 4 and 5, the user is
created using the value of "user_role" variable (this is checked using
command "openstack role assignment list --names").
PASS: Log in the system with each user created in the items 1, 3, 4 and
5 from previous step and check that the permissions are working
accordingly using a GET command ("system service-parameter-list") and a
non-GET command ("system service-parameter-delete <uuid>").
PASS: Successfully execute the LDAP playbook deleting the users created
in the steps above. Try logging into the system to assure that they are
not present.
PASS: Repeat the tests above using a Distributed Cloud (with 1 subcloud)
instead of AIO-DX. The tests should be done in both central cloud and
subcloud.

Story: 2010149
Task: 46347

Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I39e8af289433e737f8ce1a5ff7007031752c1640