Fix certificate ssl_ca cert install by dc-orch sync
This commit fixes an issue where the subcloud's kubernetes root CA
is installed as a ssl_ca in the subcloud. This is not needed and will
result in an 'Cannot install certificate with same subject' error when
dc-orch tries to synchronize it's kubernetes root CA as a ssl_ca
certificate to the subcloud.

Test Plan:

PASS: Run dcmanager subcloud add and verify that no ssl_ca certificates
      with subject 'CN=starlingx' exists in 'system certificate-list'
PASS: Bootstrap a standalone system and verify that a ssl_ca certificate
      with subject 'CN=starlingx' exists in 'system certificate-list'
PASS: Manage a subcloud and verify that it's able to synchronize certs
      without the 'Cannot install certificate with same subject' error

Partial-Bug: 1990007

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: Ia685a9a7db609de5d41e83ec4268d837da9d5010