Fix platform restore to handle missing resolv.conf
The resolv.conf file is no longer written to:
 /opt/platform/config/<ver>/resolv.conf

It was removed from that location by:
https://review.opendev.org/c/starlingx/config/+/856562

Any backup taken after that commit merged will not
contain the resolv.conf in that location, and any restore
would fail.

The resolv.conf is no longer needed at that location,
so updating the restore code to no longer try to restore it
from there.

Test Plan:
  PASS - Perform a Backup / Restore of AIO-SX Debian

Closes-Bug: #1990171

Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Id9a091d9b8d2e5e3422f8483e7dff15833efdcaa