Rename backup user images parameter
This commit will rename the old backup_user_local_registry backup
parameter to backup_user_images in order to clarify the meaning of this
operation for the users. Also, the default naming of the backup files
generated will follow the same naming pattern.

TEST PLAN
PASS Backup with new option, check if the user images backup was
     taken properly

Story: 2010117
Task: 46410
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: Id77f0bf55899a12888f6dfba81ba3fa732f5ab03