Add option for installing openldap CA cert
This change added a var in create-openldap-certificate-for-upgrade.yml
as an option to install openldap CA cert or not. This option can be
overridden by --extra-vars when the playbook is played.

In upgrade script, this option will be overridden to not install
the openldap CA cert as a temporary fix so the upgrade script won't
fail.

Test Plan:
PASS: Invoke the playbook with default value (install CA cert)
      ansible-playbook /usr/share/ansible/stx-ansible/playbooks
          /create-openldap-certificate-for-upgrade.yml
PASS: Invoke the playbook with override value "no" (not to install
      CA cert)
      ansible-playbook -e install_ca_cert=no /usr/share/ansible
          /stx-ansible/playbooks
          /create-openldap-certificate-for-upgrade.yml

Story: 2009834
Task: 46456
Signed-off-by: Andy Ning <andy.ning@windriver.com>
Change-Id: If946b80102d7128db22cbf7550f7d4736cd8a101