Debian DC - Prestage ostree repo
Update packages prestage related playbooks for Debian ostree repo
prestaging. The prestage packages step from centos is replaced with
a rsync from a depth=-1 copy of the current ostree repo on
the system controller to the subcloud in /opt/platform-backup.
The rsync is initiated from the system controller to the subcloud.
Images prestage is unchanged.

Note that support is required in miniboot.cfg to incorporate
the prestaged ostree repo during the upgrade.

Test Plan:
PASSED: Individual subcloud prestage for reinstall and upgrade
PASSED: Orchestrated prestage for reinstall and upgrade

Story: 2010120
Task: 46527

Signed-off-by: Li Zhu <li.zhu@windriver.com>
Change-Id: I03f45bbedbcb1a1c197d6d890739565a14c21f33