Debian: Fix optimized restore ldap puppet failures
When optimized backup restore on Debian is invoked,
the ldap bootstrap manifest is invoked and fails
because it requires the /etc/pki/ca-cert directory.

In a normal bootstrap, that directory is created by
another playbook.  Adding a similar setup for optimized
restore.

Note: this change is required due to a recent puppet change
https://review.opendev.org/c/starlingx/stx-puppet/+/860511

Test Plan:
  PASS: Debian: Optimized restore ansible phase completes.

Story: 2010117
Task: 45990
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Ia82140065fe8b44a343a0c39e74bec1024981a00