Install openldap CA certificate for subcloud
Since openldap is running on System Controller as a centralized
service, the signing CA certificate needs to be installed in the
subcloud in order for the clients (sssd, ldap utilies and
ldapscripts) on subcloud to access the service. This is done by
ansible bootstrap during subcloud addition.

Test Plan:
PASS: DC system controller and subcloud deployment.
PASS: On subcloud, verify openldap CA certificate is stored as
      secret "system-local-ca" in cert-manager namespace.
PASS: On subcloud after controller is unlocked, verify openldap
      CA certificate is included in trusted CA bundle:
      /etc/ssl/certs/ca-certificates.crt

Story: 2009834
Task: 46601
Signed-off-by: Andy Ning <andy.ning@windriver.com>
Change-Id: Iaccac34ab3abb1c0595e5d12d7a565151781c06b