add back code setting kubernetes_long_version
In commit f730bfd2 which was recently merged we removed what appeared
to be duplicate definitions of "kubernetes_long_version".

It turns out that these were not in fact duplicates, they are needed
when sysinv runs the upgrade-static-images.yml and
upgrade-k8s-networking.yml playbooks.  Sysinv specifies the value of
kubernetes_version, but we need to determine kubernetes_long_version.

The failure wasn't noticed immediately because the lab installs
without an error.  You have to look at the sysinv logs to see it.

For simplicity, we add back the code to generate it, but make it
conditional on not being specified yet.

Test Plan:
PASS: Manually run both playbooks that were causing problems.
PASS: Install lab with new code, ensure sysinv logs show
      no problems.

Story: 2010301
Task: 46313

Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: Ic31ddacf9d6f4c75fcc4fb5bc98c2da9e5ec3fcb