Fix ldapmodify failure during AIO-SX upgrade
During AIO-SX upgrade, ldapmodify command in puppet will fail.
The root cause is that the /etc/ldap/certs.ldif from the backup
overwrites the newer version installed by the to-release image.
This update fixed this issue by removing the certs.ldif from
restore so that the newer version is kept. This update also added
certs.ldif in the backup exclude list to decrease the size of the
backup file.

Test Plan:
PASS: AIO-SX upgrade, verify the restore upgrade data step is
      successful:
      ansible-playbook /usr/share/ansible/stx-ansible/playbooks
      /upgrade_platform.yml

Story: 2009834
Task: 46710
Signed-off-by: Andy Ning <andy.ning@windriver.com>
Change-Id: I78925f342848945dd6c50786460b7302cdaba308