Fix LDAP user password set in first login
In Debian, the text to be expected when setting the user password in the
first login is different from CentOS, so the Ansible expect module call
is updated in this commit.

Test Plan:

PASS: In an AIO-SX with this change present, create a new user following
the instructions from https://docs.starlingx.io/security/kubernetes/
manage-local-ldap-39fe3a85a528.html#manage-local-ldap-39fe3a85a528 and
successfully login through SSH using the new user.

Closes-Bug: 1995747
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I32bb055ca2409008c15a644cda248608af349f8b