Improved network setup during optimized restore
An issue noted during a previous review mentioned that "gethostip"
is not guaranteed on all systems.  It has been replaced by
grepping /etc/hosts for relevant IPs.

Ignore errors has also been removed because the issue that was
causing it to be required is not longer present on current builds.
It is unknown what was causing it.

A change also had to be made for restoring systems
that used registry.central.  Since dnsmasq is not restored
until after unlock, registry.central was unreachable.
Solution chosen is to use the platform hosts file during image
downloading sequence because it contains a
host entry for registry.central.

TEST PLAN
PASS: Optimized restore is successful (SX)
PASS: Optimized restore is successful (SX) on a subcloud

Story: 2010117
Task: 46749
Change-Id: I783fd88c6f19709c11af16013e62892502701a88